+++ loman/__init__.py
@@ -1,6 +1,6 @@
-from loman.computeengine import (
-    Computation, ComputationFactory, MapException, LoopDetectedException, NonExistentNodeException,
-    node, C, input_node, calc_node)
+from loman.computeengine import (Computation, ComputationFactory, node, C, input_node, calc_node)
+from loman.exception import (MapException, LoopDetectedException, NonExistentNodeException,
+                             CannotInsertToPlaceholderNodeException)
 from loman.consts import States, NodeTransformations
 import loman.visualization as visualization
 from loman.visualization import GraphView