+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add links parameter to Computation.add_block
 - Add keep_values parameter to Computation.add_block
 - BUGFIX: Linking a node to itself is a no-op
+- BUGFIX: Inserting to a placeholder node raises a specific exception
 
 ## [0.5.0] (2025-04-10)